Refactor some of ArimBansLib as separate repo ArimAPI
See https://github.com/A248/ArimAPI
