+++ ArimBans3/src/main/java/space/arim/bans/internal/backend/resolver/ResolverMaster.java
@@ -24,10 +24,10 @@ import java.util.UUID;
 
 import space.arim.bans.api.exception.MissingCacheException;
 import space.arim.bans.api.exception.NoGeoIpException;
-import space.arim.bans.api.exception.PlayerNotFoundException;
 import space.arim.bans.internal.Component;
 
-import space.arim.api.UUIDResolver;
+import space.arim.api.uuid.PlayerNotFoundException;
+import space.arim.api.uuid.UUIDResolver;
 import space.arim.api.util.web.GeoIpInfo;
 
 public interface ResolverMaster extends Component, UUIDResolver {
