+++ ArimBansLib/pom.xml
@@ -20,6 +20,11 @@
 	<dependencies>
 		<dependency>
 			<groupId>space.arim</groupId>
+			<artifactId>arimapi-core</artifactId>
+			<version>0.1.0-SNAPSHOT</version>
+		</dependency>
+		<dependency>
+			<groupId>space.arim</groupId>
 			<artifactId>universal-registry</artifactId>
 		</dependency>
 		<dependency>
