--- ArimBansLib/src/main/java/space/arim/api/sql/SQLPlugin.java
@@ -1,34 +1,0 @@
-/* 
- * ArimBansLib, an API for ArimBans
- * Copyright © 2019 Anand Beh <https://www.arim.space>
- * 
- * ArimBansLib is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * 
- * ArimBansLib is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License
- * along with ArimBansLib. If not, see <https://www.gnu.org/licenses/>
- * and navigate to version 3 of the GNU General Public License.
- */
-package space.arim.api.sql;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-public interface SQLPlugin {
-	
-	void executionQueries(ExecutableQuery...queries) throws SQLException;
-	
-	void executionQuery(String query, Object...parameters) throws SQLException;
-	
-	ResultSet[] selectionQueries(ExecutableQuery...queries) throws SQLException;
-	
-	ResultSet selectionQuery(String query, Object...parameters) throws SQLException;
-	
-}
