--- ArimBansLib/src/main/java/space/arim/bans/api/exception/RateLimitException.java
@@ -1,33 +1,0 @@
-/* 
- * ArimBansLib, an API for ArimBans
- * Copyright © 2019 Anand Beh <https://www.arim.space>
- * 
- * ArimBansLib is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * 
- * ArimBansLib is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License
- * along with ArimBansLib. If not, see <https://www.gnu.org/licenses/>
- * and navigate to version 3 of the GNU General Public License.
- */
-package space.arim.bans.api.exception;
-
-public class RateLimitException extends InternalAPIException {
-
-	private static final long serialVersionUID = 3081467999598825482L;
-
-	public RateLimitException(String service) {
-		super("Rate limit reached for " + service);
-	}
-
-	public RateLimitException(String service, Exception cause) {
-		super("Rate limit reached for " + service, cause);
-	}
-
-}
